refactor(cli): extract sensor gate mapping
